Fix integer overflow guard incorrectly rejecting valid datasets

+++ src/vamanabuild.c
@@ -351,8 +351,8 @@ vamanabuild(Relation heap, Relation index, IndexInfo *indexInfo)
 		goto cleanup;
 	}
 
-	if (buildstate.numVectors > 0 &&
-		buildstate.dimensions > (int) (SIZE_MAX / sizeof(float) / (size_t) buildstate.numVectors))
+	if ((size_t) buildstate.numVectors > 0 &&
+		(size_t) buildstate.dimensions > SIZE_MAX / sizeof(float) / (size_t) buildstate.numVectors)
 		ereport(ERROR,
 				(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
 				 errmsg("vector dataset too large to index "
@@ -717,8 +717,8 @@ VamanaRebuildFromTable(Relation index)
 	}
 
 	/* Flatten vector data for SVS */
-	if (numVectors > 0 &&
-		dimensions > (int) (SIZE_MAX / sizeof(float) / (size_t) numVectors))
+	if ((size_t) numVectors > 0 &&
+		(size_t) dimensions > SIZE_MAX / sizeof(float) / (size_t) numVectors)
 		ereport(ERROR,
 				(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
 				 errmsg("vector dataset too large to index "
